+++ src/com/intellij/idea/plugin/hybris/toolwindow/RemoteSolrConnectionDialog.kt
@@ -18,6 +18,9 @@
 
 package com.intellij.idea.plugin.hybris.toolwindow
 
+import com.intellij.credentialStore.CredentialAttributes
+import com.intellij.credentialStore.Credentials
+import com.intellij.ide.passwordSafe.PasswordSafe
 import com.intellij.idea.plugin.hybris.common.HybrisConstants
 import com.intellij.idea.plugin.hybris.common.utils.HybrisI18NBundleUtils
 import com.intellij.idea.plugin.hybris.notifications.Notifications
@@ -64,8 +67,7 @@ class RemoteSolrConnectionDialog(
                 port = portTextField.text
                 isSsl = sslProtocolCheckBox.isSelected
                 solrWebroot = webrootTextField.text
-                adminLogin = usernameTextField.text
-                adminPassword = String(passwordTextField.password)
+                credentials = Credentials(usernameTextField.text, String(passwordTextField.password))
                 this
             }
 
@@ -172,7 +174,7 @@ class RemoteSolrConnectionDialog(
                 label("Username:")
                 usernameTextField = textField()
                     .align(AlignX.FILL)
-                    .bindText(settings::adminLogin.toNonNullableProperty("admin"))
+                    .enabled(false)
                     .addValidationRule("Username cannot be blank.") { it.text.isNullOrBlank() }
                     .component
             }.layout(RowLayout.PARENT_GRID)
@@ -182,23 +184,6 @@ class RemoteSolrConnectionDialog(
                 passwordTextField = passwordField()
                     .align(AlignX.FILL)
                     .enabled(false)
-                    .applyToComponent {
-                        ProgressManager.getInstance().run(object : Task.Backgroundable(project, "Retrieving password", false) {
-                            private var password: String? = null
-                            override fun run(indicator: ProgressIndicator) {
-                                password = settings.adminPassword
-                                passwordTextField.text = password
-                                passwordTextField.isEnabled = true
-                            }
-                        })
-                    }
-                    .onApply {
-                        ProgressManager.getInstance().run(object : Task.Backgroundable(project, "Persisting password", false) {
-                            override fun run(indicator: ProgressIndicator) {
-                                settings.adminPassword = String(passwordTextField.password)
-                            }
-                        })
-                    }
                     .addValidationRule("Password cannot be blank.") { it.password.isEmpty() }
                     .component
             }.layout(RowLayout.PARENT_GRID)
@@ -213,6 +198,13 @@ class RemoteSolrConnectionDialog(
     override fun applyFields() {
         super.applyFields()
 
+        ProgressManager.getInstance().run(object : Task.Backgroundable(project, "Persisting credentials", false) {
+            override fun run(indicator: ProgressIndicator) {
+                val credentialAttributes = CredentialAttributes("SAP CX - ${settings.uuid}")
+                PasswordSafe.instance.set(credentialAttributes, Credentials(usernameTextField.text, String(passwordTextField.password)))
+            }
+        })
+
         // change of the scope
         if (settings.uuid != null && originalScope != settings.scope) {
             RemoteConnectionUtil.changeRemoteConnectionScope(project, settings, originalScope)
@@ -221,7 +213,20 @@ class RemoteSolrConnectionDialog(
 
     override fun createCenterPanel() = with(panel) {
         border = JBUI.Borders.empty(16)
+        loadCredentials()
         this
+    }
+
+    private fun loadCredentials() {
+        ProgressManager.getInstance().run(object : Task.Backgroundable(project, "Retrieving credentials", false) {
+            override fun run(indicator: ProgressIndicator) {
+                passwordTextField.text = settings.password
+                passwordTextField.isEnabled = true
+
+                usernameTextField.text = settings.username
+                usernameTextField.isEnabled = true
+            }
+        })
     }
 
     override fun createLeftSideActions() = arrayOf(testConnectionButton)